Add simple pipeline and step models and use runpy in app for execution
Update cinco/__init__.py to add simple Pipeline and Step models (Step
will be used as a decorator to wrap pipeline step logic). Update
start_server.py to standardize the git clone and run process a bit; now
runpy will be used to execute the pipeline script which should create
a pipeline global var with the Pipeline model which will be extracted
and each step run by this handler.